+++ CHANGELOG.md
@@ -5,6 +5,13 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 
+## [UNRELEASED]
+
+### Fixed
+- [DEPENDENCIES] Update for security fixes
+- [INTERNAL] Update textile-js to latest version 2.0.116. Fixes #75
+
+
 ## [1.3.22] - 20230702
 
 ### Fixed